+++ pkg/datastream/stream_test.go
@@ -9,7 +9,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v4/pkg/tools"
+	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v5/pkg/tools"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )