+++ pkg/edgeworkers/edgeworkers.go
@@ -4,7 +4,7 @@ package edgeworkers
 import (
 	"errors"
 
-	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v4/pkg/session"
+	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v5/pkg/session"
 )
 
 var (
@@ -24,6 +24,7 @@ type (
 		EdgeKVNamespaces
 		EdgeWorkerIDs
 		EdgeWorkerVersions
+		Groups
 		PermissionGroups
 		Properties
 		Reports