+++ pkg/papi/property.go
@@ -7,7 +7,7 @@ import (
 	"net/http"
 	"net/url"
 
-	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v4/pkg/edgegriderr"
+	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v5/pkg/edgegriderr"
 	validation "github.com/go-ozzo/ozzo-validation/v4"
 )